feature/FLOW-1392 code review changes to update pom file and reorganise the file structures.
